+++ CHANGELOG.md
@@ -5,8 +5,9 @@
 ### Added
 
 - Options to start automatically on login and start with the main window minimized
-- It is now possible to edit the match title and these edits will be used to choose file names when exporting
 - Setting to control minimize to tray
+- It is now possible to edit the match title and these edits will be used to choose file names when exporting
+- An inbox to provide indication on the current status of replays that are waiting or have failed to parse
 
 ### Fixed
 
@@ -16,6 +17,6 @@
 
 ### Changed
 
-- The bulk operations moved to a persistent toolbar. There is now a selection menu for all/none/invert.
-- The guess patched display has been relocated
+- The bulk operations have moved to a persistent toolbar. There is now a selection menu for all/none/invert.
+- The layout of the replay display has been improved